Started on tests for GetOrAddCommit operator
