+++ src/test/scala/org/codefeedr/Core/Operators/MongoDBSpec.scala
@@ -1,0 +1,31 @@
+package org.codefeedr.Core.Operators
+
+import com.typesafe.scalalogging.LazyLogging
+import org.codefeedr.Core.Clients.MongoDB.MongoDB
+import org.codefeedr.Core.LibraryServiceSpec
+import org.mongodb.scala.Completed
+import org.scalatest.Matchers
+import org.scalatest.mockito.MockitoSugar
+
+import scala.concurrent.Future
+import scala.reflect.ClassTag
+
+class MongoDBSpec extends LibraryServiceSpec with Matchers with MockitoSugar with LazyLogging {
+
+  val mongo = new MongoDB()
+
+  def ClearCollection(collectionName: String): Future[Completed] = {
+    mongo.
+      getCollection(collectionName).
+      drop().
+      toFuture()
+  }
+
+  def InsertDocument[T : ClassTag](collectionName: String, doc: T) = {
+    mongo.
+      getCollection[T](collectionName).
+      insertOne(doc).
+      toFuture()
+  }
+
+}
